Resolve some warnings.

+++ MVVMReactiveCocoa/Resource/CBStoreHouseRefreshControl/Class/CBStoreHouseRefreshControl.m
@@ -176,7 +176,7 @@ NSString *const yKey = @"y";
 
 - (CGFloat)animationProgress
 {
-    return MIN(1.f, MAX(0, fabsf(self.realContentOffsetY)/self.dropHeight));
+    return MIN(1.f, MAX(0, fabs(self.realContentOffsetY)/self.dropHeight));
 }
 
 - (CGFloat)realContentOffsetY
